Adicionar campo de mobilidade e acessibilidade e editar palcos.html pa exibir icon em função do bool

+++ festival/views.py
@@ -51,6 +51,13 @@ def editar_concerto_view(request, concerto_id):
 def apagar_concerto_view(request, concerto_id):
     concerto = get_object_or_404(Concerto, id=concerto_id)
 
-    concerto.delete()
+    if request == POST:
+        concerto.delete()
 
-    return redirect('concertos')
+    return redirect('index')
+
+def criar_concerto_view(request):
+    form = ConcertoForm()      
+
+    context = {'form': form}
+    return render(request, 'festival/criar_concerto.html', context)
